% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{classify_outcome}
\alias{classify_outcome}
\title{Competitive outcome from invasion criteria}
\usage{
classify_outcome(inv_y, inv_b)
}
\arguments{
\item{inv_y, inv_b}{Invasion criteria.}
}
\value{
Factor with levels \code{"bacteria"} (only bacteria can invade),
\code{"coexist"} (both), \code{"yeast"} (only yeast), and \code{"neither"}.
}
\description{
Competitive outcome from invasion criteria
}
