% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic.R
\name{determ_sim}
\alias{determ_sim}
\title{Simulate the deterministic one-plant model}
\usage{
determ_sim(y0, p0, times, pyr, pbr, pars = ss_params(), immigration = TRUE)
}
\arguments{
\item{y0, p0}{Initial values.}

\item{times}{Output times.}

\item{pyr, pbr}{Regional pools.}

\item{pars}{An \code{\link[=ss_params]{ss_params()}} object.}

\item{immigration}{Whether arriving pollinators carry microbes.}
}
\value{
A data frame from \code{\link[deSolve:ode]{deSolve::ode()}}.
}
\description{
Simulate the deterministic one-plant model
}
