% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Bray–Curtis dissimilarity.}
\usage{
dissimilarity(yeast, bact)
}
\arguments{
\item{yeast}{Vector of yeast abundances.}

\item{bact}{Vector of bacteria abundances.}
}
\value{
A single number indicating the mean dissimilarity across the
two vectors.
}
\description{
Bray–Curtis dissimilarity.
}
