% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dissimilarity_vector}
\alias{dissimilarity_vector}
\title{Bray–Curtis dissimilarity on grouped vectors.}
\usage{
dissimilarity_vector(yeast, bact, group_size, overall_mean = FALSE)
}
\arguments{
\item{yeast}{Vector of yeast abundances.}

\item{bact}{Vector of bacteria abundances.}

\item{group_size}{Size of groups within which dissimilarities should be
calculated. It's assumed that the vectors are
all sorted by this grouping such that every \code{group_size} elements
in the vectors belong to the same group.
The \code{yeast} and \code{bact} vectors' lengths should be divisible by
\code{group_size}.}

\item{overall_mean}{Single logical for whether to return the overall mean
dissimilarity after grouping. Defaults to \code{FALSE}.}
}
\value{
If \code{overall_mean} is \code{FALSE}, a numeric vector of length
\code{length(yeast) \%/\% group_size} with the mean dissimilarities across
the vectors for each group.
Otherwise, a single number indicating the mean dissimilarity across the
two vectors after grouping.
}
\description{
Bray–Curtis dissimilarity on grouped vectors.
}
