% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diversity}
\alias{diversity}
\title{Shannon diversity index.}
\usage{
diversity(yeast, bact, zero_threshold = 2.220446e-16)
}
\arguments{
\item{yeast}{Vector of yeast abundances.}

\item{bact}{Vector of bacteria abundances.}
}
\value{
A single number indicating the mean diversity across the
two vectors.
}
\description{
Shannon diversity index.
}
