% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diversity_vector}
\alias{diversity_vector}
\title{Shannon diversity index after summing by groups.}
\usage{
diversity_vector(yeast, bact, group_size, zero_threshold = 2.220446e-16)
}
\arguments{
\item{yeast}{Vector of yeast abundances.}

\item{bact}{Vector of bacteria abundances.}

\item{group_size}{Size of groups within which abundances should be
summed before diversities are calculated.
It's assumed that the vectors are all sorted by this grouping such
that every \code{group_size} elements in the vectors belong to the same group.
The \code{yeast} and \code{bact} vectors' lengths should be divisible by
\code{group_size}.}
}
\value{
A single number indicating the mean diversities across the
two vectors after summing species by groups.
}
\description{
Shannon diversity index after summing by groups.
}
