% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced.R
\name{joint_means}
\alias{joint_means}
\title{Means of a joint distribution over (y, p)}
\usage{
joint_means(joint)
}
\arguments{
\item{joint}{Output of \code{\link[=joint_distribution]{joint_distribution()}}.}
}
\value{
Named vector with means of y and p.
}
\description{
Means of a joint distribution over (y, p)
}
