% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_spat_wts}
\alias{make_spat_wts}
\title{Create a matrix of spatial weights from distance matrix}
\usage{
make_spat_wts(dm, m = 2)
}
\arguments{
\item{dm}{Numeric distance matrix. Must be symmetrical with zeros on
the diagonal and only contain non-negative values.
Missing values are not allowed.}

\item{m}{A single numeric indicating the exponent for spatial weighting.
Spatial weighting is \code{d^m} for each item \code{d} in \code{dm}.
The default is \code{2}.}
}
\value{
A symmetrical numeric matrix with spatial weighting between points.
}
\description{
Create a matrix of spatial weights from distance matrix
}
