% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_vcv_mat}
\alias{make_vcv_mat}
\title{Create variance-covariance matrix from distance matrix}
\usage{
make_vcv_mat(dm, q, sigma)
}
\arguments{
\item{dm}{Square, numeric distance matrix. Cannot have negative values.}

\item{q}{Single number indicating exponential distance decay constant.
Must be >= 0.}

\item{sigma}{Single number or numeric vector of the same length as the
number of rows in \code{dm}, indicating the standard deviations for
all patches (if a single numeric) or for each patch individually
(if a numeric vector).
Cannot have negative values.}
}
\value{
A square, numeric variance-covariance matrix.
}
\description{
Create variance-covariance matrix from distance matrix
}
