% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{p_mean_given_y}
\alias{p_mean_given_y}
\title{Mean pollinators on a plant given its yeast-dominated flowers}
\usage{
p_mean_given_y(y, pyr, pbr, pars = ss_params())
}
\arguments{
\item{y}{Number of yeast-dominated flowers (vectorized).}

\item{pyr, pbr}{Regional pools.}

\item{pars}{An \code{\link[=ss_params]{ss_params()}} object.}
}
\description{
Balance of pollinator immigration and emigration in the reduced model:
\code{(m_b * pbr + m * pyr) / (m_b * (n - y) + m * y)}.
}
