% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_metacomm.R
\name{plant_metacomm}
\alias{plant_metacomm}
\title{Simulate plant metacommunities.}
\usage{
plant_metacomm(
  np,
  u,
  Y0 = NULL,
  B0 = NULL,
  m = 0.1,
  d_yp = 1.1,
  d_b0 = 0.3,
  d_bp = 0.4,
  g_yp = 0.005,
  g_b0 = 0.02,
  g_bp = 0.002,
  L_0 = 0.5,
  X = 0,
  dt = 0.1,
  max_t = 500,
  no_immig = TRUE,
  open_sys = TRUE
)
}
\arguments{
\item{np}{Number of plants.}

\item{u}{Affects how strongly pollinators are attracted to plants with the
greatest proportion of palatable nectar.}

\item{Y0}{Starting yeast proportions.
If \code{NULL}, both yeast and bacteria are evenly distributed among
plants where \code{all(Y0 + B0) == 0.5}.
If \code{NULL}, \code{B0} must also be.}

\item{B0}{Starting bacteria proportions.
If \code{NULL}, both yeast and bacteria are evenly distributed among
plants where \code{all(Y0 + B0) == 0.5}.
If \code{NULL}, \code{Y0} must also be.}

\item{m}{Flower senescence.}

\item{d_yp}{Pollinator-dependent, within-plant dispersal rate for yeast.}

\item{d_b0}{Pollinator-independent, within-plant dispersal rate for bacteria.}

\item{d_bp}{Pollinator-dependent, within-plant dispersal rate for bacteria.}

\item{g_yp}{Pollinator-dependent rate of dispersal from sources other
than the focal plant for yeast.
This is overridden by the \code{no_immig} parameter by default!}

\item{g_b0}{Pollinator-independent rate of dispersal from sources other
than the focal plant for bacteria.
This is overridden by the \code{no_immig} parameter by default!}

\item{g_bp}{Pollinator-dependent rate of dispersal from sources other
than the focal plant for bacteria.
This is overridden by the \code{no_immig} parameter by default!}

\item{L_0}{Half saturation ratio for the effect of pollinators on dispersal.}

\item{X}{Attraction of pollinators to non-focal plants.}

\item{dt}{Time step for ODE solver.}

\item{max_t}{Number of days to simulate.}

\item{no_immig}{Single logical for whether to disallow immigration, which
will (if set to \code{TRUE}), override all the \verb{g_*} parameters and set them
to zero. Defaults to \code{TRUE}.}

\item{open_sys}{Single logical for whether immigration is "open" and comes
from outside the system. If \code{FALSE}, immigration for each species
is affected by their mean abundance across all plants.
If \code{TRUE}, immigration is a constant trickle from outside the landscape.
Defaults to \code{TRUE}.}
}
\value{
A data frame of yeast, bacteria, and pollinator densities at each
plant through time.
}
\description{
Simulate plant metacommunities.
}
