% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_metacomm_stoch.R
\name{plant_metacomm_stoch}
\alias{plant_metacomm_stoch}
\title{Simulate plant metacommunities with stochasticity.}
\usage{
plant_metacomm_stoch(
  np,
  u,
  Y0 = NULL,
  B0 = NULL,
  m = 0.1,
  d_yp = 1.1,
  d_b0 = 0.3,
  d_bp = 0.4,
  g_yp = 0.005,
  g_b0 = 0.02,
  g_bp = 0.002,
  L_0 = 1/np,
  X = 0,
  n_sigma = 200,
  season_len = 150,
  season_surv = 0.1,
  season_sigma = 0,
  n_reps = 100,
  dt = 0.1,
  max_t = 3000,
  closed = TRUE
)
}
\arguments{
\item{np}{Number of plants.}

\item{u}{Affects how strongly pollinators are attracted to plants with the
greatest proportion of palatable nectar.}

\item{Y0}{Starting yeast proportions.
If \code{NULL}, both yeast and bacteria are evenly distributed among
plants where \code{all(Y0 + B0) == 0.5}.
If \code{NULL}, \code{B0} must also be.}

\item{B0}{Starting bacteria proportions.
If \code{NULL}, both yeast and bacteria are evenly distributed among
plants where \code{all(Y0 + B0) == 0.5}.
If \code{NULL}, \code{Y0} must also be.}

\item{m}{Flower senescence.}

\item{d_yp}{Pollinator-dependent, within-plant dispersal rate for yeast.}

\item{d_b0}{Pollinator-independent, within-plant dispersal rate for bacteria.}

\item{d_bp}{Pollinator-dependent, within-plant dispersal rate for bacteria.}

\item{g_yp}{Pollinator-dependent rate of dispersal from sources other
than the focal plant for yeast.
This is overridden by the \code{closed} parameter by default!}

\item{g_b0}{Pollinator-independent rate of dispersal from sources other
than the focal plant for bacteria.
This is overridden by the \code{closed} parameter by default!}

\item{g_bp}{Pollinator-dependent rate of dispersal from sources other
than the focal plant for bacteria.
This is overridden by the \code{closed} parameter by default!}

\item{L_0}{Half saturation ratio for the effect of pollinators on dispersal.}

\item{X}{Attraction of pollinators to non-focal plants.}

\item{n_sigma}{Sample size of the approximated binomial distribution that
generates stochasticity at each time step.}

\item{season_len}{Length of seasons. Set to \code{NULL} to not have seasons.}

\item{season_surv}{Survival of microbes from the end of one season to
the start of the next.}

\item{season_sigma}{Standard deviation of the normal distribution used to
generate between-season variation in logit-transformed relative
species compositions.}

\item{n_reps}{Number of reps to simulate.}

\item{dt}{Time step for ODE solver.}

\item{max_t}{Number of days to simulate.}

\item{closed}{Single logical for a closed system, which will override
all the \verb{g_*} parameters and set them to zero. Defaults to TRUE.}
}
\value{
A data frame of yeast, bacteria, and pollinator densities at each
plant through time and for each repetition.
}
\description{
Simulate plant metacommunities with stochasticity.
}
