% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancy.R
\name{poisson_weights}
\alias{poisson_weights}
\title{Poisson weights for the pollinator count at senescence}
\usage{
poisson_weights(y, pyr, pbr, pars = ss_params(), normalize = TRUE)
}
\arguments{
\item{y}{Number of yeast-dominated flowers before the senescence.}

\item{pyr, pbr}{Regional pools.}

\item{pars}{An \code{\link[=ss_params]{ss_params()}} object.}

\item{normalize}{Renormalize the Poisson probabilities truncated at
\code{pmax} to sum to 1.}
}
\description{
Poisson weights for the pollinator count at senescence
}
