% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot to a PDF file using \code{cairo_pdf}}
\usage{
save_plot(fn, p, w, h, seed = NULL, ...)
}
\arguments{
\item{fn}{Filename of plot}

\item{p}{Plot object or function to create plot}

\item{w}{Width in inches}

\item{h}{Height in inches}

\item{seed}{Integer to seed RNG for consistent jittering.}

\item{...}{Other arguments to \code{cairo_pdf}}
}
\value{
\code{NULL}
}
\description{
Save a plot to a PDF file using \code{cairo_pdf}
}
