% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie.R
\name{temporal_mean}
\alias{temporal_mean}
\title{Time-weighted mean of a Gillespie path}
\usage{
temporal_mean(sim)
}
\arguments{
\item{sim}{Output of \code{\link[=gillespie_reduced]{gillespie_reduced()}}.}
}
\value{
Named vector of time-averaged \code{y}, \code{b}, \code{p}.
}
\description{
Time-weighted mean of a Gillespie path
}
