% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{update_params}
\alias{update_params}
\title{Change parameter values}
\usage{
update_params(pars, ...)
}
\arguments{
\item{pars}{An \code{ss_params} object.}

\item{...}{Named values to replace.}
}
\value{
An \code{ss_params} object.
}
\description{
Change parameter values
}
